backend/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
python-decouple
django-extensions
celery
redis
drf-yasg
django-filter
djangorestframework-simplejwt
django-allauth
django-rest-auth
drf-spectacular
channels
channels-redis
django-tenant-schemas
django-tenants
psycopg2-binary
django-guardian
django-role-permissions
wechatpayv3
alipay-sdk-python
pycryptodome
qrcode
escpos
qrcode[pil]
reportlab
django-qr-code